CSHARP-2812: Code review changes.
